modified init call params

+++ Processor/Processor.h
@@ -291,7 +291,7 @@ public:
 	~spdz_ext_ifc();
 
     void * ext_lib_handle;
-    int (*ext_init)(const int pid, const char * field, const size_t offline_size);
+    int (*ext_init)(const int pid, const char * field, const int offline_size);
     int (*ext_start_open)(const size_t share_count, const char ** shares);
     int (*ext_stop_open)(void *);
     int (*ext_term)(void *);
